Rewrite progress report header to use a table view
* Eliminates a lot of AutoLayout code that may be causing crashes
* Replaces confusing view mode switching with a popover from the title
* Add a "totals" view that shows points in each category
* Add 3D Touch to the header
* Increase logging to find the logout bug

+++ iOS/break/break/break/LoopMail/LoopMailComposeViewController.swift
@@ -204,8 +204,8 @@ class LoopMailComposeViewController: UIViewController, UITableViewDataSource, UI
 				return
 		}
 		let convertedKeyboardEndFrame = composeTableView.convert(keyboardEndFrame, from: composeTableView.window)
-		let rawAnimationCurve = ((userInfo[UIKeyboardAnimationCurveUserInfoKey] as? NSNumber)?.uintValue ?? 0)
-		let animationCurve = UIViewAnimationOptions(rawValue: UInt(rawAnimationCurve))
+		let rawAnimationCurve = (userInfo[UIKeyboardAnimationCurveUserInfoKey] as? NSNumber)?.uintValue ?? 0
+		let animationCurve = UIViewAnimationOptions(rawValue: rawAnimationCurve)
 		composeTableView.contentInset = UIEdgeInsets(top: composeTableView.contentInset.top, left: composeTableView.contentInset.left, bottom: max(composeTableView.bounds.maxY - convertedKeyboardEndFrame.minY, tabBarController?.tabBar.frame.height ?? 0), right: composeTableView.contentInset.right)
 		composeTableView.scrollIndicatorInsets = UIEdgeInsets(top: composeTableView.scrollIndicatorInsets.top, left: composeTableView.scrollIndicatorInsets.left, bottom: max(composeTableView.bounds.maxY - convertedKeyboardEndFrame.minY, tabBarController?.tabBar.frame.height ?? 0), right: composeTableView.scrollIndicatorInsets.right)
 		composeTableView.flashScrollIndicators()
